requirements-dev.txt
# Development dependencies
pytest>=7.0
pytest-cov>=4.0
black>=22.0
flake8>=5.0
mypy>=1.0
python-dotenv>=1.0.0
twine>=4.0.0
build>=1.0.0
